Create Utils class and define getUserNickname() function to use instead of getNickname() from the Users API function and fix bug with deleteComments()
