+++ portfolio/src/main/java/com/google/sps/data/UserAuthenticationData.java
@@ -14,20 +14,20 @@
 
 package com.google.sps.data;
 
-/** A comment from user input. */
-public final class Comment {
+/**
+ * The data used to authenticate a user.
+ */
+public final class UserAuthenticationData {
 
-  private final long id;
-  private final String username;
-  private final String content;
-  private final String imageUrl;
-  private final long timestamp;
+  private final boolean loggedIn;
+  private final String authenticationUrl;
+  private final String email;
+  private final String nickname; // User-provided display name.
 
-  public Comment(long id, String username, String content, String imageUrl, long timestamp) {
-    this.id = id;
-    this.username = username;
-    this.content = content;
-    this.imageUrl = imageUrl;
-    this.timestamp = timestamp;
+  public UserAuthenticationData(boolean loggedIn, String authenticationUrl, String email, String nickname) {
+    this.loggedIn = loggedIn;
+    this.authenticationUrl = authenticationUrl;
+    this.email = email;
+    this.nickname = nickname;
   }
 }
